requirements.txt
fastapi
Flask
gymnasium
jinja2
numpy
paho-mqtt
scikit-learn
sqlalchemy
tensorflow
tensorflow_federated
uvicorn[standard]
